requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Brotli==1.1.0

certifi==2023.11.17
charset-normalizer==3.3.2

dj-database-url==2.1.0
Django==4.0
django-cors-headers==4.3.1

djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

gunicorn==21.2.0

idna==3.6

mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9

PyJWT==2.8.0

pytz==2023.3.post1
requests==2.31.0

sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
